requirements.txt
numpy==1.23.4
pandas==1.5.1
scikit_learn==1.1.3
streamlit==1.14.1
plotly
altair
Pillow
joblib
